[cogito-mobile] Home screen
