+++ workspaces/cogito-mobile/App.js
@@ -1,19 +1,3 @@
-import React from 'react'
-import { StyleSheet, Text, View } from 'react-native'
+import { Home } from './Home'
 
-const App = () => (
-  <View style={styles.container}>
-    <Text>Open up App.js to start working on your app!</Text>
-  </View>
-)
-
-export default App
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center'
-  }
-})
+export default Home
