+++ workspaces/cogito-mobile/App.test.js
@@ -1,15 +1,8 @@
-/**
- * @format
- * @lint-ignore-every XPLATJSCOPYRIGHT1
- */
-
-import 'react-native'
 import React from 'react'
+import { render } from 'react-native-testing-library'
 import App from './App'
 
-// Note: test renderer must be required after react-native.
-import renderer from 'react-test-renderer'
-
-it('renders correctly', () => {
-  renderer.create(<App />)
+it('show the home screen by default', () => {
+  const { queryByText } = render(<App />)
+  expect(queryByText('I am.')).not.toBeNull()
 })
