+++ workspaces/cogito-mobile/Home.js
@@ -1,0 +1,17 @@
+import React from 'react'
+import { StyleSheet, Text, View } from 'react-native'
+
+export const Home = () => (
+  <View style={styles.container}>
+    <Text>I am.</Text>
+  </View>
+)
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center'
+  }
+})
